Debug Studio 20

+++ SharedCode/BasicDisplayVisitor.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 void BasicDisplayVisitor::visit_ImageFile(ImageFile* filePtr) {
 
-	int board_size = sqrt(filePtr->getSize());
+	int board_size = (int)sqrt(filePtr->getSize());
 
 	for (int i = 0; i < board_size; ++i) {
